requirements.txt
pyserial==3.3
websocket_client==0.46.0
psutil==5.2.2
subprocess32==3.2.7
wsaccel==0.6.2